[CLIPPER-118] Create per-model and per-application metrics (#128)
* Added application and model metrics

* Code compiles and unit tests pass

* fixed segfault

* all metrics getting registered

* changed get_app endpoints to POST and fixed broken url

* reverted get_app changes

* Addressed review comments

* fixed merge error

* Addressed another round of review comments

+++ src/libclipper/test/input_test.cpp
@@ -132,7 +132,7 @@ TEST(InputSerializationTests, IntSerialization) {
 
   long* raw_input_header_size =
       reinterpret_cast<long*>(input_header_size.data());
-  ASSERT_EQ(raw_input_header_size[0], input_header.size());
+  ASSERT_EQ(static_cast<size_t>(raw_input_header_size[0]), input_header.size());
   long* raw_input_content_size =
       reinterpret_cast<long*>(input_content_size.data());
   ASSERT_EQ(raw_input_content_size[0], input_content.size());
